Safari-611.1.10/LayoutTests/fast/canvas/imagedata-contains-uint8clampedarray-expected.txt

Tests that ImageData contains a Uint8ClampedArray.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS imageData.data.constructor is Uint8ClampedArray
PASS successfullyParsed is true

TEST COMPLETE
